+++ src/lib/Components/Bidding/Screens/ConfirmBid.tsx
@@ -1,3 +1,4 @@
+import { isEmpty } from "lodash"
 import React from "react"
 import { NativeModules, NavigatorIOS, View, ViewProperties } from "react-native"
 import { commitMutation, createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
@@ -176,7 +177,7 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConfirmBidState
     })
 
     commitMutation(this.props.relay.environment, {
-      onCompleted: () => this.createBidderPosition(),
+      onCompleted: (_, errors) => (isEmpty(errors) ? this.createBidderPosition() : this.presentErrorResult(errors)),
       onError: this.presentErrorResult.bind(this),
       mutation: creditCardMutation,
       variables: {
@@ -189,7 +190,8 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConfirmBidState
 
   createBidderPosition() {
     commitMutation(this.props.relay.environment, {
-      onCompleted: (results, errors) => this.verifyBidPosition(results, errors),
+      onCompleted: (results, errors) =>
+        isEmpty(errors) ? this.verifyBidPosition(results) : this.presentErrorResult(errors),
       onError: this.presentErrorResult.bind(this),
       mutation: bidderPositionMutation,
       variables: {
@@ -202,10 +204,10 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConfirmBidState
     })
   }
 
-  verifyBidPosition(results, errors) {
+  verifyBidPosition(results) {
     const { result } = results.createBidderPosition
 
-    if (!errors && result.status === "SUCCESS") {
+    if (result.status === "SUCCESS") {
       this.bidPlacedSuccessfully(result.position.id)
     } else {
       this.presentBidResult(result)